framer fix: transfition beforeChilder
- I'm trying to improve animation performance because it seems on some devices that the opening animations of the modal sometimes are laggy.

I still did not figure out were the problem relies, but I'm thinking that it has something to do with overlaping animations.

When you create an habit, the modal closes, triggering the exit animations. At the same time, because of the store update, the habits list layout animation also gets triggered and on some devices this process is laggy.