Feat:Habits sorting without global state mutations
- Created 2 new hooks to keep the logic and implementation details inside the store and only expose the hooks to the component/pages that need them.

- useHabits( sortCriteria ) -> by sortCriteria order if any or by default order
- useHabit( id )

Also extracted the input state and setter of the global state because it didn't make sense to be there vecause it was just a local state that can be perfectly managed at the component level. Also avoiding in this way unnesesary code bloat at the store.

Added new action to clear all habits at the store.