refactor: check and update habits months data
- Implemented a store actions that checks if each habit has the needed data and adds it if necesary.

Example: Lets say that one of our users haven't used the app for a while, but someday he decides to open it up again.

When he opens up the app again and goes to any of the habits that he have had created before, he will find that the last month he will see is the last month that he used the app on. Because of that he won't be able to update his progress on those months if any.

I implemented the checkAndUpdateHabits store action to solve this problem.

Basicaly what it does is that, when executed, maps the habits array and checks wether the habit has or not the needed data, this means, if the habit has in the months property the current month as the last one. If not it will add each month since the last one registered to the current month